Rename Snapshot -> Dataset to clarify intent
Originally created this class to represent snapshots of reservoir data, but it looks like it could be more multi-purpose and you might want to use it by supplying a dataset rather than only exporting from a reservoir. Not sure this name is perfect, but it is better. Naming is hard. :-)